moved StatusDetailFragment
