+++ app/src/main/java/com/freshdigitable/udonroad/TimelineContainerSwitcher.java
@@ -29,6 +29,7 @@ import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 
+import com.freshdigitable.udonroad.detail.StatusDetailFragment;
 import com.freshdigitable.udonroad.ffab.IndicatableFFAB;
 
 import java.util.ArrayList;
